refactor chunk buffer generation into seperate class
