+++ src/render/StructureRenderer.ts
@@ -1,14 +1,14 @@
-import { mat4, vec3 } from 'gl-matrix'
-import type { Identifier, PlacedBlock, StructureProvider } from '../core/index.js'
-import { BlockPos, BlockState, Direction } from '../core/index.js'
+import type { vec3 } from 'gl-matrix'
+import { mat4 } from 'gl-matrix'
+import type { Identifier, StructureProvider } from '../core/index.js'
+import { BlockState } from '../core/index.js'
 import type { BlockDefinitionProvider } from './BlockDefinition.js'
 import type { BlockModelProvider } from './BlockModel.js'
-import type { RenderBuffers } from './Renderer.js'
+import { ChunkBuilder } from './ChunkBuilder.js'
 import { Renderer } from './Renderer.js'
 import { ShaderProgram } from './ShaderProgram.js'
-import { SpecialRenderer, SpecialRenderers } from './SpecialRenderer.js'
 import type { TextureAtlasProvider } from './TextureAtlas.js'
-import { mergeFloat32Arrays, transformVectors } from './Util.js'
+import { createBuffer } from './Util.js'
 
 const vsColor = `
   attribute vec4 vertPos;
@@ -58,20 +58,8 @@ const fsGrid = `
   }
 `
 
-interface ChunkRenderBuffers extends RenderBuffers {
-	blockPos: WebGLBuffer,
-}
 
-interface RenderChunk {
-	positions: Float32Array[],
-	textureCoordinates: number[],
-	tintColors: number[],
-	normals: number[],
-	blockPositions: number[],
-	indices: number[],
-	indexOffset: number,
-	buffer?: ChunkRenderBuffers,
-}
+
 
 type GridBuffers = {
 	position: WebGLBuffer,
@@ -101,18 +89,16 @@ interface StructureRendererOptions {
 }
 
 export class StructureRenderer extends Renderer {
-	private readonly chunkSize: number
-
 	private readonly gridShaderProgram: WebGLProgram
 	private readonly colorShaderProgram: WebGLProgram
-	
-	private chunks: RenderChunk[][][] = []
 
 	private gridBuffers: GridBuffers
 	private readonly outlineBuffers: GridBuffers
 	private invisibleBlockBuffers: GridBuffers | undefined
 	private readonly atlasTexture: WebGLTexture
 	private readonly useInvisibleBlockBuffer: boolean
+
+	private readonly chunkBuilder: ChunkBuilder
 
 	constructor(
 		gl: WebGLRenderingContext,
@@ -122,7 +108,10 @@ export class StructureRenderer extends Renderer {
 	) {
 		super(gl)
 
-		this.chunkSize = options?.chunkSize ?? 16
+		const chunkSize = options?.chunkSize ?? 16
+
+		this.chunkBuilder = new ChunkBuilder(gl, structure, resources, chunkSize)
+
 		if (options?.facesPerBuffer){
 			console.warn('[deepslate renderer warning]: facesPerBuffer option has been removed in favor of chunkSize')
 		}
@@ -131,7 +120,6 @@ export class StructureRenderer extends Renderer {
 		this.gridShaderProgram = new ShaderProgram(gl, vsGrid, fsGrid).getProgram()
 		this.colorShaderProgram = new ShaderProgram(gl, vsColor, fsColor).getProgram()
 
-		this.updateStructureBuffers()
 		this.gridBuffers = this.getGridBuffers()
 		this.outlineBuffers = this.getOutlineBuffers()
 		this.invisibleBlockBuffers = this.getInvisibleBlockBuffers()
@@ -140,162 +128,13 @@ export class StructureRenderer extends Renderer {
 
 	public setStructure(structure: StructureProvider) {
 		this.structure = structure
-		this.updateStructureBuffers()
+		this.chunkBuilder.setStructure(structure)
 		this.gridBuffers = this.getGridBuffers()
 		this.invisibleBlockBuffers = this.getInvisibleBlockBuffers()
 	}
-	private getChunk(chunkPos: vec3): RenderChunk{
-		const x = Math.abs(chunkPos[0]) * 2 + (chunkPos[0] < 0 ? 1 : 0)
-		const y = Math.abs(chunkPos[1]) * 2 + (chunkPos[1] < 0 ? 1 : 0)
-		const z = Math.abs(chunkPos[2]) * 2 + (chunkPos[2] < 0 ? 1 : 0)
-
-		if (!this.chunks[x])
-			this.chunks[x] = []
-		if (!this.chunks[x][y])
-			this.chunks[x][y] = []
-		if (!this.chunks[x][y][z])
-			this.chunks[x][y][z] = {
-				positions: [],
-				textureCoordinates: [],
-				tintColors: [],
-				blockPositions: [],
-				normals: [],
-				indices: [],
-				indexOffset: 0,
-			}
-
-		return this.chunks[x][y][z]
-	}  
 
 	public updateStructureBuffers(chunkPositions?: vec3[]): void {
-
-		const pushBuffers = (buffers: any, pos: vec3, chunk: RenderChunk) => {
-			const t = mat4.create()
-			mat4.translate(t, t, pos)
-			transformVectors(buffers.position, t)
-
-			chunk.positions.push(buffers.position)
-			chunk.textureCoordinates.push(...buffers.texCoord)
-			chunk.tintColors.push(...buffers.tintColor)
-			for (let i = 0; i < buffers.position.length; i += 12) {
-				const a = vec3.fromValues(buffers.position[i], buffers.position[i + 1], buffers.position[i + 2])
-				const b = vec3.fromValues(buffers.position[i + 3], buffers.position[i + 4], buffers.position[i + 5])
-				const c = vec3.fromValues(buffers.position[i + 6], buffers.position[i + 7], buffers.position[i + 8])
-				vec3.subtract(b, b, a)
-				vec3.subtract(c, c, a)
-				vec3.cross(b, b, c)
-				vec3.normalize(b, b)
-				chunk.normals.push(...b, ...b, ...b, ...b)
-			}
-			for (let i = 0; i < buffers.texCoord.length / 2; i += 1) {
-				chunk.blockPositions.push(...pos)
-			}
-			chunk.indices.push(...buffers.index)
-			chunk.indexOffset += buffers.texCoord.length / 2
-		}
-
-		const resetChunk = (chunk: RenderChunk) => {
-			chunk.positions = []
-			chunk.textureCoordinates = []
-			chunk.tintColors = []
-			chunk.blockPositions = []
-			chunk.normals = []
-			chunk.indices = []
-			chunk.indexOffset = 0
-		}
-
-		const refreshBuffer = (chunk: RenderChunk) => {
-			if (chunk.buffer) {
-				this.updateBuffer(chunk.buffer.position, this.gl.ARRAY_BUFFER, mergeFloat32Arrays(...chunk.positions))
-				this.updateBuffer(chunk.buffer.texCoord, this.gl.ARRAY_BUFFER, new Float32Array(chunk.textureCoordinates)),
-				this.updateBuffer(chunk.buffer.tintColor, this.gl.ARRAY_BUFFER, new Float32Array(chunk.tintColors)),
-				this.updateBuffer(chunk.buffer.normal, this.gl.ARRAY_BUFFER, new Float32Array(chunk.normals)),
-				this.updateBuffer(chunk.buffer.blockPos, this.gl.ARRAY_BUFFER, new Float32Array(chunk.blockPositions)),
-				this.updateBuffer(chunk.buffer.index, this.gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(chunk.indices)),
-				chunk.buffer.length = chunk.indices.length
-			} else {
-				chunk.buffer = {
-					position: this.createBuffer(this.gl.ARRAY_BUFFER, mergeFloat32Arrays(...chunk.positions)),
-					texCoord: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(chunk.textureCoordinates)),
-					tintColor: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(chunk.tintColors)),
-					blockPos: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(chunk.blockPositions)),
-					normal: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(chunk.normals)),
-					index: this.createBuffer(this.gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(chunk.indices)),
-					length: chunk.indices.length,
-				}
-			}
-		}
-
-		if (!chunkPositions){
-			this.chunks.forEach(x => x.forEach(y => y.forEach(chunk => {
-				resetChunk(chunk)
-			})))
-		} else { 
-			chunkPositions.forEach(chunkPos => {
-				const chunk = this.getChunk(chunkPos)
-				resetChunk(chunk)
-			})
-		}
-
-		let buffers
-		for (const b of this.structure.getBlocks()) {
-			const blockName = b.state.getName()
-			const blockProps = b.state.getProperties()
-			const defaultProps = this.resources.getDefaultBlockProperties(blockName) ?? {}
-			Object.entries(defaultProps).forEach(([k, v]) => {
-				if (!blockProps[k]) blockProps[k] = v
-			})
-
-			const chunkPos: vec3 = [Math.floor(b.pos[0]/this.chunkSize), Math.floor(b.pos[1]/this.chunkSize), Math.floor(b.pos[2]/this.chunkSize)]
-
-			if (chunkPositions && !chunkPositions.some(pos => vec3.equals(pos, chunkPos)))
-				continue
-
-			const chunk = this.getChunk(chunkPos)
-
-			try {
-				const blockDefinition = this.resources.getBlockDefinition(blockName)
-				const cull = {
-					up: this.needsCull(b, Direction.UP),
-					down: this.needsCull(b, Direction.DOWN),
-					west: this.needsCull(b, Direction.WEST),
-					east: this.needsCull(b, Direction.EAST),
-					north: this.needsCull(b, Direction.NORTH),
-					south: this.needsCull(b, Direction.SOUTH),
-				}
-				if (blockDefinition) {
-					buffers = blockDefinition.getBuffers(blockName, blockProps, this.resources, this.resources, chunk.indexOffset, cull)
-					pushBuffers(buffers, b.pos, chunk)
-				}
-				if (SpecialRenderers.has(blockName.toString())) {
-					buffers = SpecialRenderer[blockName.toString()](chunk.indexOffset, blockProps, this.resources, cull)
-					pushBuffers(buffers, b.pos, chunk)
-				}
-			} catch(e) {
-				console.error(`Error rendering block ${blockName}`, e)
-			}
-		}
-
-		if (!chunkPositions){
-			this.chunks.forEach(x => x.forEach(y => y.forEach(chunk => {
-				refreshBuffer(chunk)
-			})))
-		} else {
-			chunkPositions.forEach(chunkPos => {
-				const chunk = this.getChunk(chunkPos)
-				refreshBuffer(chunk)
-			})
-		}
-	}
-
-	private needsCull(block: PlacedBlock, dir: Direction) {
-		const neighbor = this.structure.getBlock(BlockPos.towards(block.pos, dir))?.state
-		if (!neighbor) return false
-		if (this.resources.getBlockFlags(neighbor.getName())?.opaque) {
-			return !(dir === Direction.UP && block.state.isFluid())
-		} else {
-			return block.state.isFluid() && neighbor.isFluid()
-		}
+		this.chunkBuilder.updateStructureBuffers(chunkPositions)
 	}
 
 	private getGridBuffers(): GridBuffers {
@@ -324,8 +163,8 @@ export class StructureRenderer extends Renderer {
 		for (let i = 0; i < 8 + X + Z; i += 1) color.push(0.8, 0.8, 0.8, 0.8, 0.8, 0.8)
 
 		return {
-			position: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(position)),
-			color: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(color)),
+			position: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(position)),
+			color: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(color)),
 			length: position.length / 3,
 		}
 	}
@@ -337,8 +176,8 @@ export class StructureRenderer extends Renderer {
 		this.addCube(position, color, [1, 1, 1], [0, 0, 0], [1, 1, 1])
 
 		return {
-			position: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(position)),
-			color: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(color)),
+			position: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(position)),
+			color: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(color)),
 			length: position.length / 3,
 		}
 	}
@@ -369,8 +208,8 @@ export class StructureRenderer extends Renderer {
 		}
 
 		return {
-			position: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(position)),
-			color: this.createBuffer(this.gl.ARRAY_BUFFER, new Float32Array(color)),
+			position: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(position)),
+			color: createBuffer(this.gl, this.gl.ARRAY_BUFFER, new Float32Array(color)),
 			length: position.length / 3,
 		}
 	}
@@ -421,13 +260,8 @@ export class StructureRenderer extends Renderer {
 		this.setTexture(this.atlasTexture)
 		this.prepareDraw(viewMatrix)
 
-		this.chunks.forEach(x => {
-			x.forEach(y => {
-				y.forEach(chunk => {
-					if (!chunk.buffer) return
-					this.drawBuffers(chunk.buffer)
-				})
-			})
+		this.chunkBuilder.getBuffers().forEach(buffer => {
+			this.drawBuffers(buffer)
 		})
 	}
 
@@ -435,17 +269,12 @@ export class StructureRenderer extends Renderer {
 		this.setShader(this.colorShaderProgram)
 		this.prepareDraw(viewMatrix)
 
-		this.chunks.forEach(x => {
-			x.forEach(y => {
-				y.forEach(chunk => {
-					if (!chunk.buffer) return
-					this.setVertexAttr('vertPos', 3, chunk.buffer.position)
-					this.setVertexAttr('blockPos', 3, chunk.buffer.blockPos)
-					this.gl.bindBuffer(this.gl.ELEMENT_ARRAY_BUFFER, chunk.buffer.index)
+		this.chunkBuilder.getBuffers().forEach(buffer => {
+			this.setVertexAttr('vertPos', 3, buffer.position)
+			this.setVertexAttr('blockPos', 3, buffer.blockPos)
+			this.gl.bindBuffer(this.gl.ELEMENT_ARRAY_BUFFER, buffer.index)
 
-					this.gl.drawElements(this.gl.TRIANGLES, chunk.buffer.length, this.gl.UNSIGNED_SHORT, 0)
-				})
-			})
+			this.gl.drawElements(this.gl.TRIANGLES, buffer.length, this.gl.UNSIGNED_SHORT, 0)
 		})
 	}
