+++ src/render/index.ts
@@ -1,6 +1,7 @@
 export * from './BlockColors.js'
 export * from './BlockDefinition.js'
 export * from './BlockModel.js'
+export * from './ChunkBuilder.js'
 export * from './Cull.js'
 export * from './ItemColors.js'
 export * from './ItemRenderer.js'
@@ -8,3 +9,4 @@ export * from './Renderer.js'
 export * from './ShaderProgram.js'
 export * from './StructureRenderer.js'
 export * from './TextureAtlas.js'
+
